fix: :bug: change isAdmin definition method
